# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BreakBPM API — accounts, passes, and game persistence
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /passes/redeem:
    post:
      operationId: redeemDiscountCode
      tags: [passes]
      summary: Redeem a discount code (issues a pass)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DiscountRedeemInput" }
      responses:
        "200":
          description: Redeemed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RedeemResult" }
...
  /passes/checkout:
    post:
      operationId: createPassCheckout
      tags: [passes]
      summary: Begin a pass purchase — returns a checkout URL + opaque token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PassCheckoutInput" }
      responses:
        "200":
          description: Checkout created (or rejected if no provider)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckoutResult" }
  /passes/verify:
    post:
      operationId: verifyPassCheckout
      tags: [passes]
      summary: Verify a checkout token and grant the pass on success
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyCheckoutInput" }
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerifyResult" }

  # TODO(remove-before-launch): dev-only free Lifetime grant endpoint.
  # Gated server-side by DEV_FREE_UPGRADE_ENABLED flag; returns 404 when off.
...
components:
  schemas:
...
    PassSummary:
      type: object
      required: [kind, startedAt, expiresAt, isLifetime]
      properties:
        kind: { type: string, enum: [day, year, lifetime] }
        startedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        isLifetime: { type: boolean }
...
    DiscountRedeemInput:
      type: object
      required: [code]
      properties:
        code: { type: string, minLength: 1, maxLength: 64 }
    RedeemResult:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean }
        message: { type: string }
        pass: { $ref: "#/components/schemas/PassSummary" }
    PassCheckoutInput:
      type: object
      required: [kind]
      properties:
        # Only one-time passes are purchasable. The former one-time "year" pass
        # is replaced by the Yearly subscription, so it is no longer accepted
        # here (legacy year-pass rows still resolve entitlement, but no new ones
        # can be bought).
        kind: { type: string, enum: [day, lifetime] }
    CheckoutResult:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean }
        message: { type: string }
        # Provider-issued opaque token. Hand this to /passes/verify to
        # complete the purchase. Absent when success=false.
        opaqueToken: { type: string }
        # Optional URL to redirect the user to (e.g. Stripe Checkout).
        checkoutUrl: { type: string }
    VerifyCheckoutInput:
      type: object
      required: [opaqueToken]
      properties:
        opaqueToken: { type: string, minLength: 1, maxLength: 512 }
    VerifyResult:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean }
        message: { type: string }
        pass: { $ref: "#/components/schemas/PassSummary" }

    # TODO(remove-before-launch): paired with /passes/dev-grant-lifetime.
...